[refactor] clean up
