--- tests/dataflow/aie/gpt2/kernels/lut_softmax.cc
@@ -1,563 +1,0 @@
-/*
- * Copyright Allo authors. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#include <aie_api/aie.hpp>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <type_traits>
-
-#define NOCPP
-
-alignas(aie::vector_decl_align) unsigned char m_inv_lut[128] = {
-    0,   126, 124, 122, 120, 118, 117, 115, 113, 111, 109, 108, 106, 104, 103,
-    101, 100, 98,  96,  95,  93,  92,  90,  89,  88,  86,  85,  83,  82,  81,
-    79,  78,  77,  76,  74,  73,  72,  71,  69,  68,  67,  66,  65,  64,  63,
-    61,  60,  59,  58,  57,  56,  55,  54,  53,  52,  51,  50,  49,  48,  47,
-    46,  45,  44,  44,  43,  42,  41,  40,  39,  38,  37,  37,  36,  35,  34,
-    33,  33,  32,  31,  30,  30,  29,  28,  27,  27,  26,  25,  24,  24,  23,
-    22,  22,  21,  20,  20,  19,  18,  18,  17,  16,  16,  15,  14,  14,  13,
-    13,  12,  11,  11,  10,  10,  9,   9,   8,   7,   7,   6,   6,   5,   5,
-    4,   4,   3,   3,   2,   2,   1,   1};
-
-// The tables _ab and _cd are copies of each other
-// Also, each table has a copy of data seperated by 128-bit
-// This effectively creates 4 copies of the table (located in 4 (128-bit) banks)
-// This allows the user to use the "gather" read feature and read 4 values from
-// 4 different banks at once Tables entries are in BF16
-//
-#if __AIE_ARCH__ == 20
-alignas(aie::vector_decl_align) int16 softmax_ilut_ab[512] = {
-    16256, 16430, 16620, 16801, 16986, 17172, 17354, 17545, 16256, 16430, 16620,
-    16801, 16986, 17172, 17354, 17545, 17722, 17917, 18092, 18282, 18463, 18648,
-    18835, 19016, 17722, 17917, 18092, 18282, 18463, 18648, 18835, 19016, 19208,
-    19384, 19578, 19754, 19943, 20125, 20310, 20497, 19208, 19384, 19578, 19754,
-    19943, 20125, 20310, 20497, 20677, 20870, 21046, 21240, 21416, 21605, 21788,
-    21971, 20677, 20870, 21046, 21240, 21416, 21605, 21788, 21971, 22160, 22339,
-    22533, 22708, 22901, 23079, 23266, 23450, 22160, 22339, 22533, 22708, 22901,
-    23079, 23266, 23450, 23633, 23822, 24001, 24195, 24370, 24562, 24741, 24928,
-    23633, 23822, 24001, 24195, 24370, 24562, 24741, 24928, 25112, 25295, 25485,
-    25663, 25858, 26032, 26224, 26403, 25112, 25295, 25485, 25663, 25858, 26032,
-    26224, 26403, 26589, 26774, 26957, 27147, 27325, 27520, 27695, 27885, 26589,
-    26774, 26957, 27147, 27325, 27520, 27695, 27885, 28065, 28251, 28437, 28618,
-    28809, 28987, 29182, 29357, 28065, 28251, 28437, 28618, 28809, 28987, 29182,
-    29357, 29547, 29727, 29913, 30099, 30280, 30472, 30649, 30843, 29547, 29727,
-    29913, 30099, 30280, 30472, 30649, 30843, 31019, 31208, 31390, 31574, 31762,
-    31942, 32135, 32311, 31019, 31208, 31390, 31574, 31762, 31942, 32135, 32311,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     1,     3,     9,     24,    0,     0,
-    0,     0,     1,     3,     9,     24,    66,    179,   372,   550,   737,
-    921,   1104,  1293,  66,    179,   372,   550,   737,   921,   1104,  1293,
-    1472,  1666,  1841,  2033,  2212,  2399,  2583,  2766,  1472,  1666,  1841,
-    2033,  2212,  2399,  2583,  2766,  2956,  3134,  3329,  3503,  3694,  3874,
-    4060,  4246,  2956,  3134,  3329,  3503,  3694,  3874,  4060,  4246,  4427,
-    4618,  4796,  4991,  5165,  5356,  5536,  5722,  4427,  4618,  4796,  4991,
-    5165,  5356,  5536,  5722,  5908,  6089,  6281,  6458,  6652,  6828,  7017,
-    7198,  5908,  6089,  6281,  6458,  6652,  6828,  7017,  7198,  7383,  7570,
-    7751,  7943,  8120,  8314,  8490,  8679,  7383,  7570,  7751,  7943,  8120,
-    8314,  8490,  8679,  8861,  9045,  9233,  9413,  9606,  9782,  9975,  10152,
-    8861,  9045,  9233,  9413,  9606,  9782,  9975,  10152, 10340, 10523, 10707,
-    10895, 11075, 11268, 11444, 11636, 10340, 10523, 10707, 10895, 11075, 11268,
-    11444, 11636, 11814, 12002, 12185, 12368, 12558, 12737, 12931, 13106, 11814,
-    12002, 12185, 12368, 12558, 12737, 12931, 13106, 13298, 13476, 13663, 13848,
-    14030, 14220, 14398, 14593, 13298, 13476, 13663, 13848, 14030, 14220, 14398,
-    14593, 14768, 14959, 15138, 15325, 15510, 15692, 15883, 16060, 14768, 14959,
-    15138, 15325, 15510, 15692, 15883, 16060};
-
-alignas(aie::vector_decl_align) int16 softmax_ilut_cd[512] = {
-    16256, 16430, 16620, 16801, 16986, 17172, 17354, 17545, 16256, 16430, 16620,
-    16801, 16986, 17172, 17354, 17545, 17722, 17917, 18092, 18282, 18463, 18648,
-    18835, 19016, 17722, 17917, 18092, 18282, 18463, 18648, 18835, 19016, 19208,
-    19384, 19578, 19754, 19943, 20125, 20310, 20497, 19208, 19384, 19578, 19754,
-    19943, 20125, 20310, 20497, 20677, 20870, 21046, 21240, 21416, 21605, 21788,
-    21971, 20677, 20870, 21046, 21240, 21416, 21605, 21788, 21971, 22160, 22339,
-    22533, 22708, 22901, 23079, 23266, 23450, 22160, 22339, 22533, 22708, 22901,
-    23079, 23266, 23450, 23633, 23822, 24001, 24195, 24370, 24562, 24741, 24928,
-    23633, 23822, 24001, 24195, 24370, 24562, 24741, 24928, 25112, 25295, 25485,
-    25663, 25858, 26032, 26224, 26403, 25112, 25295, 25485, 25663, 25858, 26032,
-    26224, 26403, 26589, 26774, 26957, 27147, 27325, 27520, 27695, 27885, 26589,
-    26774, 26957, 27147, 27325, 27520, 27695, 27885, 28065, 28251, 28437, 28618,
-    28809, 28987, 29182, 29357, 28065, 28251, 28437, 28618, 28809, 28987, 29182,
-    29357, 29547, 29727, 29913, 30099, 30280, 30472, 30649, 30843, 29547, 29727,
-    29913, 30099, 30280, 30472, 30649, 30843, 31019, 31208, 31390, 31574, 31762,
-    31942, 32135, 32311, 31019, 31208, 31390, 31574, 31762, 31942, 32135, 32311,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     1,     3,     9,     24,    0,     0,
-    0,     0,     1,     3,     9,     24,    66,    179,   372,   550,   737,
-    921,   1104,  1293,  66,    179,   372,   550,   737,   921,   1104,  1293,
-    1472,  1666,  1841,  2033,  2212,  2399,  2583,  2766,  1472,  1666,  1841,
-    2033,  2212,  2399,  2583,  2766,  2956,  3134,  3329,  3503,  3694,  3874,
-    4060,  4246,  2956,  3134,  3329,  3503,  3694,  3874,  4060,  4246,  4427,
-    4618,  4796,  4991,  5165,  5356,  5536,  5722,  4427,  4618,  4796,  4991,
-    5165,  5356,  5536,  5722,  5908,  6089,  6281,  6458,  6652,  6828,  7017,
-    7198,  5908,  6089,  6281,  6458,  6652,  6828,  7017,  7198,  7383,  7570,
-    7751,  7943,  8120,  8314,  8490,  8679,  7383,  7570,  7751,  7943,  8120,
-    8314,  8490,  8679,  8861,  9045,  9233,  9413,  9606,  9782,  9975,  10152,
-    8861,  9045,  9233,  9413,  9606,  9782,  9975,  10152, 10340, 10523, 10707,
-    10895, 11075, 11268, 11444, 11636, 10340, 10523, 10707, 10895, 11075, 11268,
-    11444, 11636, 11814, 12002, 12185, 12368, 12558, 12737, 12931, 13106, 11814,
-    12002, 12185, 12368, 12558, 12737, 12931, 13106, 13298, 13476, 13663, 13848,
-    14030, 14220, 14398, 14593, 13298, 13476, 13663, 13848, 14030, 14220, 14398,
-    14593, 14768, 14959, 15138, 15325, 15510, 15692, 15883, 16060, 14768, 14959,
-    15138, 15325, 15510, 15692, 15883, 16060};
-
-alignas(aie::vector_decl_align) int16 softmax_flut_ab[512] = {
-    16256, 16257, 16257, 16258, 16258, 16259, 16259, 16260, 16256, 16257, 16257,
-    16258, 16258, 16259, 16259, 16260, 16260, 16261, 16261, 16262, 16262, 16263,
-    16263, 16264, 16260, 16261, 16261, 16262, 16262, 16263, 16263, 16264, 16264,
-    16265, 16265, 16266, 16266, 16267, 16267, 16268, 16264, 16265, 16265, 16266,
-    16266, 16267, 16267, 16268, 16269, 16269, 16270, 16270, 16271, 16271, 16272,
-    16272, 16269, 16269, 16270, 16270, 16271, 16271, 16272, 16272, 16273, 16274,
-    16274, 16275, 16275, 16276, 16276, 16277, 16273, 16274, 16274, 16275, 16275,
-    16276, 16276, 16277, 16278, 16278, 16279, 16279, 16280, 16281, 16281, 16282,
-    16278, 16278, 16279, 16279, 16280, 16281, 16281, 16282, 16282, 16283, 16284,
-    16284, 16285, 16285, 16286, 16287, 16282, 16283, 16284, 16284, 16285, 16285,
-    16286, 16287, 16287, 16288, 16289, 16289, 16290, 16290, 16291, 16292, 16287,
-    16288, 16289, 16289, 16290, 16290, 16291, 16292, 16292, 16293, 16294, 16294,
-    16295, 16296, 16296, 16297, 16292, 16293, 16294, 16294, 16295, 16296, 16296,
-    16297, 16298, 16298, 16299, 16300, 16300, 16301, 16302, 16302, 16298, 16298,
-    16299, 16300, 16300, 16301, 16302, 16302, 16303, 16304, 16304, 16305, 16306,
-    16306, 16307, 16308, 16303, 16304, 16304, 16305, 16306, 16306, 16307, 16308,
-    16309, 16309, 16310, 16311, 16311, 16312, 16313, 16314, 16309, 16309, 16310,
-    16311, 16311, 16312, 16313, 16314, 16314, 16315, 16316, 16316, 16317, 16318,
-    16319, 16319, 16314, 16315, 16316, 16316, 16317, 16318, 16319, 16319, 16320,
-    16321, 16322, 16322, 16323, 16324, 16325, 16325, 16320, 16321, 16322, 16322,
-    16323, 16324, 16325, 16325, 16326, 16327, 16328, 16329, 16329, 16330, 16331,
-    16332, 16326, 16327, 16328, 16329, 16329, 16330, 16331, 16332, 16333, 16333,
-    16334, 16335, 16336, 16337, 16337, 16338, 16333, 16333, 16334, 16335, 16336,
-    16337, 16337, 16338, 16339, 16340, 16341, 16342, 16342, 16343, 16344, 16345,
-    16339, 16340, 16341, 16342, 16342, 16343, 16344, 16345, 16346, 16347, 16347,
-    16348, 16349, 16350, 16351, 16352, 16346, 16347, 16347, 16348, 16349, 16350,
-    16351, 16352, 16353, 16354, 16354, 16355, 16356, 16357, 16358, 16359, 16353,
-    16354, 16354, 16355, 16356, 16357, 16358, 16359, 16360, 16361, 16362, 16363,
-    16363, 16364, 16365, 16366, 16360, 16361, 16362, 16363, 16363, 16364, 16365,
-    16366, 16367, 16368, 16369, 16370, 16371, 16372, 16373, 16374, 16367, 16368,
-    16369, 16370, 16371, 16372, 16373, 16374, 16375, 16376, 16377, 16378, 16379,
-    16380, 16381, 16382, 16375, 16376, 16377, 16378, 16379, 16380, 16381, 16382,
-    16383, 16384, 16384, 16385, 16385, 16386, 16386, 16387, 16383, 16384, 16384,
-    16385, 16385, 16386, 16386, 16387, 16387, 16388, 16388, 16389, 16389, 16390,
-    16390, 16391, 16387, 16388, 16388, 16389, 16389, 16390, 16390, 16391, 16391,
-    16392, 16393, 16393, 16394, 16394, 16395, 16395, 16391, 16392, 16393, 16393,
-    16394, 16394, 16395, 16395, 16396, 16396, 16397, 16397, 16398, 16399, 16399,
-    16400, 16396, 16396, 16397, 16397, 16398, 16399, 16399, 16400, 16400, 16401,
-    16401, 16402, 16402, 16403, 16404, 16404, 16400, 16401, 16401, 16402, 16402,
-    16403, 16404, 16404, 16405, 16405, 16406, 16407, 16407, 16408, 16408, 16409,
-    16405, 16405, 16406, 16407, 16407, 16408, 16408, 16409, 16410, 16410, 16411,
-    16411, 16412, 16413, 16413, 16414, 16410, 16410, 16411, 16411, 16412, 16413,
-    16413, 16414, 16414, 16415, 16416, 16416, 16417, 16418, 16418, 16419, 16414,
-    16415, 16416, 16416, 16417, 16418, 16418, 16419, 16419, 16420, 16421, 16421,
-    16422, 16423, 16423, 16424, 16419, 16420, 16421, 16421, 16422, 16423, 16423,
-    16424, 16425, 16425, 16426, 16427, 16427, 16428, 16429, 16429, 16425, 16425,
-    16426, 16427, 16427, 16428, 16429, 16429};
-
-alignas(aie::vector_decl_align) int16 softmax_flut_cd[512] = {
-    16256, 16257, 16257, 16258, 16258, 16259, 16259, 16260, 16256, 16257, 16257,
-    16258, 16258, 16259, 16259, 16260, 16260, 16261, 16261, 16262, 16262, 16263,
-    16263, 16264, 16260, 16261, 16261, 16262, 16262, 16263, 16263, 16264, 16264,
-    16265, 16265, 16266, 16266, 16267, 16267, 16268, 16264, 16265, 16265, 16266,
-    16266, 16267, 16267, 16268, 16269, 16269, 16270, 16270, 16271, 16271, 16272,
-    16272, 16269, 16269, 16270, 16270, 16271, 16271, 16272, 16272, 16273, 16274,
-    16274, 16275, 16275, 16276, 16276, 16277, 16273, 16274, 16274, 16275, 16275,
-    16276, 16276, 16277, 16278, 16278, 16279, 16279, 16280, 16281, 16281, 16282,
-    16278, 16278, 16279, 16279, 16280, 16281, 16281, 16282, 16282, 16283, 16284,
-    16284, 16285, 16285, 16286, 16287, 16282, 16283, 16284, 16284, 16285, 16285,
-    16286, 16287, 16287, 16288, 16289, 16289, 16290, 16290, 16291, 16292, 16287,
-    16288, 16289, 16289, 16290, 16290, 16291, 16292, 16292, 16293, 16294, 16294,
-    16295, 16296, 16296, 16297, 16292, 16293, 16294, 16294, 16295, 16296, 16296,
-    16297, 16298, 16298, 16299, 16300, 16300, 16301, 16302, 16302, 16298, 16298,
-    16299, 16300, 16300, 16301, 16302, 16302, 16303, 16304, 16304, 16305, 16306,
-    16306, 16307, 16308, 16303, 16304, 16304, 16305, 16306, 16306, 16307, 16308,
-    16309, 16309, 16310, 16311, 16311, 16312, 16313, 16314, 16309, 16309, 16310,
-    16311, 16311, 16312, 16313, 16314, 16314, 16315, 16316, 16316, 16317, 16318,
-    16319, 16319, 16314, 16315, 16316, 16316, 16317, 16318, 16319, 16319, 16320,
-    16321, 16322, 16322, 16323, 16324, 16325, 16325, 16320, 16321, 16322, 16322,
-    16323, 16324, 16325, 16325, 16326, 16327, 16328, 16329, 16329, 16330, 16331,
-    16332, 16326, 16327, 16328, 16329, 16329, 16330, 16331, 16332, 16333, 16333,
-    16334, 16335, 16336, 16337, 16337, 16338, 16333, 16333, 16334, 16335, 16336,
-    16337, 16337, 16338, 16339, 16340, 16341, 16342, 16342, 16343, 16344, 16345,
-    16339, 16340, 16341, 16342, 16342, 16343, 16344, 16345, 16346, 16347, 16347,
-    16348, 16349, 16350, 16351, 16352, 16346, 16347, 16347, 16348, 16349, 16350,
-    16351, 16352, 16353, 16354, 16354, 16355, 16356, 16357, 16358, 16359, 16353,
-    16354, 16354, 16355, 16356, 16357, 16358, 16359, 16360, 16361, 16362, 16363,
-    16363, 16364, 16365, 16366, 16360, 16361, 16362, 16363, 16363, 16364, 16365,
-    16366, 16367, 16368, 16369, 16370, 16371, 16372, 16373, 16374, 16367, 16368,
-    16369, 16370, 16371, 16372, 16373, 16374, 16375, 16376, 16377, 16378, 16379,
-    16380, 16381, 16382, 16375, 16376, 16377, 16378, 16379, 16380, 16381, 16382,
-    16383, 16384, 16384, 16385, 16385, 16386, 16386, 16387, 16383, 16384, 16384,
-    16385, 16385, 16386, 16386, 16387, 16387, 16388, 16388, 16389, 16389, 16390,
-    16390, 16391, 16387, 16388, 16388, 16389, 16389, 16390, 16390, 16391, 16391,
-    16392, 16393, 16393, 16394, 16394, 16395, 16395, 16391, 16392, 16393, 16393,
-    16394, 16394, 16395, 16395, 16396, 16396, 16397, 16397, 16398, 16399, 16399,
-    16400, 16396, 16396, 16397, 16397, 16398, 16399, 16399, 16400, 16400, 16401,
-    16401, 16402, 16402, 16403, 16404, 16404, 16400, 16401, 16401, 16402, 16402,
-    16403, 16404, 16404, 16405, 16405, 16406, 16407, 16407, 16408, 16408, 16409,
-    16405, 16405, 16406, 16407, 16407, 16408, 16408, 16409, 16410, 16410, 16411,
-    16411, 16412, 16413, 16413, 16414, 16410, 16410, 16411, 16411, 16412, 16413,
-    16413, 16414, 16414, 16415, 16416, 16416, 16417, 16418, 16418, 16419, 16414,
-    16415, 16416, 16416, 16417, 16418, 16418, 16419, 16419, 16420, 16421, 16421,
-    16422, 16423, 16423, 16424, 16419, 16420, 16421, 16421, 16422, 16423, 16423,
-    16424, 16425, 16425, 16426, 16427, 16427, 16428, 16429, 16429, 16425, 16425,
-    16426, 16427, 16427, 16428, 16429, 16429};
-#else
-alignas(aie::vector_decl_align) int16 softmax_ilut_ab[512] = {
-    16256, 16430, 16620, 16801, 16986, 17172, 17354, 17545, 17722, 17917, 18092,
-    18282, 18463, 18648, 18835, 19016, 16256, 16430, 16620, 16801, 16986, 17172,
-    17354, 17545, 17722, 17917, 18092, 18282, 18463, 18648, 18835, 19016, 19208,
-    19384, 19578, 19754, 19943, 20125, 20310, 20497, 20677, 20870, 21046, 21240,
-    21416, 21605, 21788, 21971, 19208, 19384, 19578, 19754, 19943, 20125, 20310,
-    20497, 20677, 20870, 21046, 21240, 21416, 21605, 21788, 21971, 22160, 22339,
-    22533, 22708, 22901, 23079, 23266, 23450, 23633, 23822, 24001, 24195, 24370,
-    24562, 24741, 24928, 22160, 22339, 22533, 22708, 22901, 23079, 23266, 23450,
-    23633, 23822, 24001, 24195, 24370, 24562, 24741, 24928, 25112, 25295, 25485,
-    25663, 25858, 26032, 26224, 26403, 26589, 26774, 26957, 27147, 27325, 27520,
-    27695, 27885, 25112, 25295, 25485, 25663, 25858, 26032, 26224, 26403, 26589,
-    26774, 26957, 27147, 27325, 27520, 27695, 27885, 28065, 28251, 28437, 28618,
-    28809, 28987, 29182, 29357, 29547, 29727, 29913, 30099, 30280, 30472, 30649,
-    30843, 28065, 28251, 28437, 28618, 28809, 28987, 29182, 29357, 29547, 29727,
-    29913, 30099, 30280, 30472, 30649, 30843, 31019, 31208, 31390, 31574, 31762,
-    31942, 32135, 32311, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    31019, 31208, 31390, 31574, 31762, 31942, 32135, 32311, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     1,     3,     9,     24,    66,    179,
-    372,   550,   737,   921,   1104,  1293,  0,     0,     0,     0,     1,
-    3,     9,     24,    66,    179,   372,   550,   737,   921,   1104,  1293,
-    1472,  1666,  1841,  2033,  2212,  2399,  2583,  2766,  2956,  3134,  3329,
-    3503,  3694,  3874,  4060,  4246,  1472,  1666,  1841,  2033,  2212,  2399,
-    2583,  2766,  2956,  3134,  3329,  3503,  3694,  3874,  4060,  4246,  4427,
-    4618,  4796,  4991,  5165,  5356,  5536,  5722,  5908,  6089,  6281,  6458,
-    6652,  6828,  7017,  7198,  4427,  4618,  4796,  4991,  5165,  5356,  5536,
-    5722,  5908,  6089,  6281,  6458,  6652,  6828,  7017,  7198,  7383,  7570,
-    7751,  7943,  8120,  8314,  8490,  8679,  8861,  9045,  9233,  9413,  9606,
-    9782,  9975,  10152, 7383,  7570,  7751,  7943,  8120,  8314,  8490,  8679,
-    8861,  9045,  9233,  9413,  9606,  9782,  9975,  10152, 10340, 10523, 10707,
-    10895, 11075, 11268, 11444, 11636, 11814, 12002, 12185, 12368, 12558, 12737,
-    12931, 13106, 10340, 10523, 10707, 10895, 11075, 11268, 11444, 11636, 11814,
-    12002, 12185, 12368, 12558, 12737, 12931, 13106, 13298, 13476, 13663, 13848,
-    14030, 14220, 14398, 14593, 14768, 14959, 15138, 15325, 15510, 15692, 15883,
-    16060, 13298, 13476, 13663, 13848, 14030, 14220, 14398, 14593, 14768, 14959,
-    15138, 15325, 15510, 15692, 15883, 16060};
-
-alignas(aie::vector_decl_align) int16 softmax_ilut_cd[512] = {
-    16256, 16430, 16620, 16801, 16986, 17172, 17354, 17545, 17722, 17917, 18092,
-    18282, 18463, 18648, 18835, 19016, 16256, 16430, 16620, 16801, 16986, 17172,
-    17354, 17545, 17722, 17917, 18092, 18282, 18463, 18648, 18835, 19016, 19208,
-    19384, 19578, 19754, 19943, 20125, 20310, 20497, 20677, 20870, 21046, 21240,
-    21416, 21605, 21788, 21971, 19208, 19384, 19578, 19754, 19943, 20125, 20310,
-    20497, 20677, 20870, 21046, 21240, 21416, 21605, 21788, 21971, 22160, 22339,
-    22533, 22708, 22901, 23079, 23266, 23450, 23633, 23822, 24001, 24195, 24370,
-    24562, 24741, 24928, 22160, 22339, 22533, 22708, 22901, 23079, 23266, 23450,
-    23633, 23822, 24001, 24195, 24370, 24562, 24741, 24928, 25112, 25295, 25485,
-    25663, 25858, 26032, 26224, 26403, 26589, 26774, 26957, 27147, 27325, 27520,
-    27695, 27885, 25112, 25295, 25485, 25663, 25858, 26032, 26224, 26403, 26589,
-    26774, 26957, 27147, 27325, 27520, 27695, 27885, 28065, 28251, 28437, 28618,
-    28809, 28987, 29182, 29357, 29547, 29727, 29913, 30099, 30280, 30472, 30649,
-    30843, 28065, 28251, 28437, 28618, 28809, 28987, 29182, 29357, 29547, 29727,
-    29913, 30099, 30280, 30472, 30649, 30843, 31019, 31208, 31390, 31574, 31762,
-    31942, 32135, 32311, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    31019, 31208, 31390, 31574, 31762, 31942, 32135, 32311, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505, 32505,
-    32505, 32505, 32505, 0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     0,     0,     0,     0,     0,     0,
-    0,     0,     0,     0,     0,     1,     3,     9,     24,    66,    179,
-    372,   550,   737,   921,   1104,  1293,  0,     0,     0,     0,     1,
-    3,     9,     24,    66,    179,   372,   550,   737,   921,   1104,  1293,
-    1472,  1666,  1841,  2033,  2212,  2399,  2583,  2766,  2956,  3134,  3329,
-    3503,  3694,  3874,  4060,  4246,  1472,  1666,  1841,  2033,  2212,  2399,
-    2583,  2766,  2956,  3134,  3329,  3503,  3694,  3874,  4060,  4246,  4427,
-    4618,  4796,  4991,  5165,  5356,  5536,  5722,  5908,  6089,  6281,  6458,
-    6652,  6828,  7017,  7198,  4427,  4618,  4796,  4991,  5165,  5356,  5536,
-    5722,  5908,  6089,  6281,  6458,  6652,  6828,  7017,  7198,  7383,  7570,
-    7751,  7943,  8120,  8314,  8490,  8679,  8861,  9045,  9233,  9413,  9606,
-    9782,  9975,  10152, 7383,  7570,  7751,  7943,  8120,  8314,  8490,  8679,
-    8861,  9045,  9233,  9413,  9606,  9782,  9975,  10152, 10340, 10523, 10707,
-    10895, 11075, 11268, 11444, 11636, 11814, 12002, 12185, 12368, 12558, 12737,
-    12931, 13106, 10340, 10523, 10707, 10895, 11075, 11268, 11444, 11636, 11814,
-    12002, 12185, 12368, 12558, 12737, 12931, 13106, 13298, 13476, 13663, 13848,
-    14030, 14220, 14398, 14593, 14768, 14959, 15138, 15325, 15510, 15692, 15883,
-    16060, 13298, 13476, 13663, 13848, 14030, 14220, 14398, 14593, 14768, 14959,
-    15138, 15325, 15510, 15692, 15883, 16060};
-
-alignas(aie::vector_decl_align) int16 softmax_flut_ab[512] = {
-    16256, 16257, 16257, 16258, 16258, 16259, 16259, 16260, 16260, 16261, 16261,
-    16262, 16262, 16263, 16263, 16264, 16256, 16257, 16257, 16258, 16258, 16259,
-    16259, 16260, 16260, 16261, 16261, 16262, 16262, 16263, 16263, 16264, 16264,
-    16265, 16265, 16266, 16266, 16267, 16267, 16268, 16269, 16269, 16270, 16270,
-    16271, 16271, 16272, 16272, 16264, 16265, 16265, 16266, 16266, 16267, 16267,
-    16268, 16269, 16269, 16270, 16270, 16271, 16271, 16272, 16272, 16273, 16274,
-    16274, 16275, 16275, 16276, 16276, 16277, 16278, 16278, 16279, 16279, 16280,
-    16281, 16281, 16282, 16273, 16274, 16274, 16275, 16275, 16276, 16276, 16277,
-    16278, 16278, 16279, 16279, 16280, 16281, 16281, 16282, 16282, 16283, 16284,
-    16284, 16285, 16285, 16286, 16287, 16287, 16288, 16289, 16289, 16290, 16290,
-    16291, 16292, 16282, 16283, 16284, 16284, 16285, 16285, 16286, 16287, 16287,
-    16288, 16289, 16289, 16290, 16290, 16291, 16292, 16292, 16293, 16294, 16294,
-    16295, 16296, 16296, 16297, 16298, 16298, 16299, 16300, 16300, 16301, 16302,
-    16302, 16292, 16293, 16294, 16294, 16295, 16296, 16296, 16297, 16298, 16298,
-    16299, 16300, 16300, 16301, 16302, 16302, 16303, 16304, 16304, 16305, 16306,
-    16306, 16307, 16308, 16309, 16309, 16310, 16311, 16311, 16312, 16313, 16314,
-    16303, 16304, 16304, 16305, 16306, 16306, 16307, 16308, 16309, 16309, 16310,
-    16311, 16311, 16312, 16313, 16314, 16314, 16315, 16316, 16316, 16317, 16318,
-    16319, 16319, 16320, 16321, 16322, 16322, 16323, 16324, 16325, 16325, 16314,
-    16315, 16316, 16316, 16317, 16318, 16319, 16319, 16320, 16321, 16322, 16322,
-    16323, 16324, 16325, 16325, 16326, 16327, 16328, 16329, 16329, 16330, 16331,
-    16332, 16333, 16333, 16334, 16335, 16336, 16337, 16337, 16338, 16326, 16327,
-    16328, 16329, 16329, 16330, 16331, 16332, 16333, 16333, 16334, 16335, 16336,
-    16337, 16337, 16338, 16339, 16340, 16341, 16342, 16342, 16343, 16344, 16345,
-    16346, 16347, 16347, 16348, 16349, 16350, 16351, 16352, 16339, 16340, 16341,
-    16342, 16342, 16343, 16344, 16345, 16346, 16347, 16347, 16348, 16349, 16350,
-    16351, 16352, 16353, 16354, 16354, 16355, 16356, 16357, 16358, 16359, 16360,
-    16361, 16362, 16363, 16363, 16364, 16365, 16366, 16353, 16354, 16354, 16355,
-    16356, 16357, 16358, 16359, 16360, 16361, 16362, 16363, 16363, 16364, 16365,
-    16366, 16367, 16368, 16369, 16370, 16371, 16372, 16373, 16374, 16375, 16376,
-    16377, 16378, 16379, 16380, 16381, 16382, 16367, 16368, 16369, 16370, 16371,
-    16372, 16373, 16374, 16375, 16376, 16377, 16378, 16379, 16380, 16381, 16382,
-    16383, 16384, 16384, 16385, 16385, 16386, 16386, 16387, 16387, 16388, 16388,
-    16389, 16389, 16390, 16390, 16391, 16383, 16384, 16384, 16385, 16385, 16386,
-    16386, 16387, 16387, 16388, 16388, 16389, 16389, 16390, 16390, 16391, 16391,
-    16392, 16393, 16393, 16394, 16394, 16395, 16395, 16396, 16396, 16397, 16397,
-    16398, 16399, 16399, 16400, 16391, 16392, 16393, 16393, 16394, 16394, 16395,
-    16395, 16396, 16396, 16397, 16397, 16398, 16399, 16399, 16400, 16400, 16401,
-    16401, 16402, 16402, 16403, 16404, 16404, 16405, 16405, 16406, 16407, 16407,
-    16408, 16408, 16409, 16400, 16401, 16401, 16402, 16402, 16403, 16404, 16404,
-    16405, 16405, 16406, 16407, 16407, 16408, 16408, 16409, 16410, 16410, 16411,
-    16411, 16412, 16413, 16413, 16414, 16414, 16415, 16416, 16416, 16417, 16418,
-    16418, 16419, 16410, 16410, 16411, 16411, 16412, 16413, 16413, 16414, 16414,
-    16415, 16416, 16416, 16417, 16418, 16418, 16419, 16419, 16420, 16421, 16421,
-    16422, 16423, 16423, 16424, 16425, 16425, 16426, 16427, 16427, 16428, 16429,
-    16429, 16419, 16420, 16421, 16421, 16422, 16423, 16423, 16424, 16425, 16425,
-    16426, 16427, 16427, 16428, 16429, 16429};
-
-alignas(aie::vector_decl_align) int16 softmax_flut_cd[512] = {
-    16256, 16257, 16257, 16258, 16258, 16259, 16259, 16260, 16260, 16261, 16261,
-    16262, 16262, 16263, 16263, 16264, 16256, 16257, 16257, 16258, 16258, 16259,
-    16259, 16260, 16260, 16261, 16261, 16262, 16262, 16263, 16263, 16264, 16264,
-    16265, 16265, 16266, 16266, 16267, 16267, 16268, 16269, 16269, 16270, 16270,
-    16271, 16271, 16272, 16272, 16264, 16265, 16265, 16266, 16266, 16267, 16267,
-    16268, 16269, 16269, 16270, 16270, 16271, 16271, 16272, 16272, 16273, 16274,
-    16274, 16275, 16275, 16276, 16276, 16277, 16278, 16278, 16279, 16279, 16280,
-    16281, 16281, 16282, 16273, 16274, 16274, 16275, 16275, 16276, 16276, 16277,
-    16278, 16278, 16279, 16279, 16280, 16281, 16281, 16282, 16282, 16283, 16284,
-    16284, 16285, 16285, 16286, 16287, 16287, 16288, 16289, 16289, 16290, 16290,
-    16291, 16292, 16282, 16283, 16284, 16284, 16285, 16285, 16286, 16287, 16287,
-    16288, 16289, 16289, 16290, 16290, 16291, 16292, 16292, 16293, 16294, 16294,
-    16295, 16296, 16296, 16297, 16298, 16298, 16299, 16300, 16300, 16301, 16302,
-    16302, 16292, 16293, 16294, 16294, 16295, 16296, 16296, 16297, 16298, 16298,
-    16299, 16300, 16300, 16301, 16302, 16302, 16303, 16304, 16304, 16305, 16306,
-    16306, 16307, 16308, 16309, 16309, 16310, 16311, 16311, 16312, 16313, 16314,
-    16303, 16304, 16304, 16305, 16306, 16306, 16307, 16308, 16309, 16309, 16310,
-    16311, 16311, 16312, 16313, 16314, 16314, 16315, 16316, 16316, 16317, 16318,
-    16319, 16319, 16320, 16321, 16322, 16322, 16323, 16324, 16325, 16325, 16314,
-    16315, 16316, 16316, 16317, 16318, 16319, 16319, 16320, 16321, 16322, 16322,
-    16323, 16324, 16325, 16325, 16326, 16327, 16328, 16329, 16329, 16330, 16331,
-    16332, 16333, 16333, 16334, 16335, 16336, 16337, 16337, 16338, 16326, 16327,
-    16328, 16329, 16329, 16330, 16331, 16332, 16333, 16333, 16334, 16335, 16336,
-    16337, 16337, 16338, 16339, 16340, 16341, 16342, 16342, 16343, 16344, 16345,
-    16346, 16347, 16347, 16348, 16349, 16350, 16351, 16352, 16339, 16340, 16341,
-    16342, 16342, 16343, 16344, 16345, 16346, 16347, 16347, 16348, 16349, 16350,
-    16351, 16352, 16353, 16354, 16354, 16355, 16356, 16357, 16358, 16359, 16360,
-    16361, 16362, 16363, 16363, 16364, 16365, 16366, 16353, 16354, 16354, 16355,
-    16356, 16357, 16358, 16359, 16360, 16361, 16362, 16363, 16363, 16364, 16365,
-    16366, 16367, 16368, 16369, 16370, 16371, 16372, 16373, 16374, 16375, 16376,
-    16377, 16378, 16379, 16380, 16381, 16382, 16367, 16368, 16369, 16370, 16371,
-    16372, 16373, 16374, 16375, 16376, 16377, 16378, 16379, 16380, 16381, 16382,
-    16383, 16384, 16384, 16385, 16385, 16386, 16386, 16387, 16387, 16388, 16388,
-    16389, 16389, 16390, 16390, 16391, 16383, 16384, 16384, 16385, 16385, 16386,
-    16386, 16387, 16387, 16388, 16388, 16389, 16389, 16390, 16390, 16391, 16391,
-    16392, 16393, 16393, 16394, 16394, 16395, 16395, 16396, 16396, 16397, 16397,
-    16398, 16399, 16399, 16400, 16391, 16392, 16393, 16393, 16394, 16394, 16395,
-    16395, 16396, 16396, 16397, 16397, 16398, 16399, 16399, 16400, 16400, 16401,
-    16401, 16402, 16402, 16403, 16404, 16404, 16405, 16405, 16406, 16407, 16407,
-    16408, 16408, 16409, 16400, 16401, 16401, 16402, 16402, 16403, 16404, 16404,
-    16405, 16405, 16406, 16407, 16407, 16408, 16408, 16409, 16410, 16410, 16411,
-    16411, 16412, 16413, 16413, 16414, 16414, 16415, 16416, 16416, 16417, 16418,
-    16418, 16419, 16410, 16410, 16411, 16411, 16412, 16413, 16413, 16414, 16414,
-    16415, 16416, 16416, 16417, 16418, 16418, 16419, 16419, 16420, 16421, 16421,
-    16422, 16423, 16423, 16424, 16425, 16425, 16426, 16427, 16427, 16428, 16429,
-    16429, 16419, 16420, 16421, 16421, 16422, 16423, 16423, 16424, 16425, 16425,
-    16426, 16427, 16427, 16428, 16429, 16429};
-#endif
-
-template <unsigned VecLen, typename VecIteratorIn, typename VecIteratorOut>
-float exp_bf16(int num_elems, VecIteratorIn &in, VecIteratorOut &out) {
-  bfloat16 __aie_dm_resource_a *ilut_ab =
-      (bfloat16 __aie_dm_resource_a *)softmax_ilut_ab;
-  bfloat16 __aie_dm_resource_b *ilut_cd =
-      (bfloat16 __aie_dm_resource_b *)softmax_ilut_cd;
-  bfloat16 __aie_dm_resource_a *flut_ab =
-      (bfloat16 __aie_dm_resource_a *)softmax_flut_ab;
-  bfloat16 __aie_dm_resource_b *flut_cd =
-      (bfloat16 __aie_dm_resource_b *)softmax_flut_cd;
-  using lut_type = aie::lut<4, bfloat16, bfloat16>;
-  const int LUT_elems = 256;
-  const int step_i = 8;
-  const int step_f = 0;
-
-  constexpr int SM_SCALE_FAC =
-      8; // Use 8-bit fractional part for LUTs when converting from bfloat16 to
-         // int, adjust any input scale factor using this.
-
-  const int elem_iters =
-      num_elems / VecLen +
-      (num_elems % VecLen != 0); // number of iterations need to be performed
-  aie::vector<bfloat16, VecLen> I_val_vec, F_val_vec, res0, input_bf16;
-  aie::accum<accfloat, VecLen> exp_val_accum;
-  aie::accum<accfloat, VecLen> exp_val_accum_shift;
-  exp_val_accum = aie::zeros<accfloat, VecLen>();
-  // Maximum value computation
-  aie::accum<accfloat, VecLen> acc0;
-  aie::vector<int16, VecLen> input;
-  aie::vector<int16, 2 * VecLen> input0;
-
-  lut_type lut_i(LUT_elems, ilut_ab, ilut_cd);
-  lut_type lut_f(LUT_elems, flut_ab, flut_cd);
-  aie::parallel_lookup<uint16, lut_type, aie::lut_oor_policy::truncate>
-      lookup_i(lut_i, step_i);
-  aie::parallel_lookup<uint16, lut_type, aie::lut_oor_policy::truncate>
-      lookup_f(lut_f, step_f);
-  aie::accum<accfloat, VecLen> exp_val;
-
-  for (int i = 0; i < elem_iters; i++)
-    chess_prepare_for_pipelining chess_loop_range(4, ) {
-      aie::vector<bfloat16, VecLen> input_org = aie::load_v<VecLen>(in);
-      in += VecLen;
-      acc0.from_vector(input_org, 0);
-      input_bf16 = to_v16bfloat16(acc0);
-      input0 = v32int16(bfloat16_to_int(input_bf16, SM_SCALE_FAC));
-#ifndef SM_USE_MSB
-      input = filter_even(input0);
-#else
-      input = filter_odd(input0);
-#endif
-
-      I_val_vec = lookup_i.fetch(input.template cast_to<uint16>());
-      F_val_vec = lookup_f.fetch(input.template cast_to<uint16>());
-      exp_val = aie::mul(I_val_vec, F_val_vec);
-      exp_val_accum = add(exp_val_accum, exp_val);
-      aie::store_v(out, exp_val.template to_vector<bfloat16>());
-      out += VecLen;
-    }
-  // Variant not using emulated FP32 for the mul reduce, off by +/- 1 in final
-  // result and 10 cycles slower
-  aie::vector<float, VecLen> reduce = exp_val_accum.template to_vector<float>();
-  float res = aie::reduce_add(reduce);
-  return res;
-}
-
-template <int L>
-void init_softmax(bfloat16 *__restrict max_logit,
-                  bfloat16 *__restrict sum_exp) {
-  // max_logit = np.full((L, 1), -np.inf)
-  // sum_exp = np.zeros((L, 1))
-  constexpr int vec_factor =
-      256 / (sizeof(bfloat16) * 8); // one 256 bit store unit
-  static_assert(L % vec_factor == 0);
-  const bfloat16 neg_inf = bfloat16(-std::numeric_limits<float>::infinity());
-  const aie::vector<bfloat16, vec_factor> neg_infs =
-      aie::broadcast<bfloat16, vec_factor>(neg_inf);
-  const aie::vector<bfloat16, vec_factor> zeros =
-      aie::zeros<bfloat16, vec_factor>();
-  for (int iter = 0; iter < L; iter += vec_factor) {
-    aie::store_v(max_logit, neg_infs);
-    max_logit += vec_factor;
-    aie::store_v(sum_exp, zeros);
-    sum_exp += vec_factor;
-  }
-}
-
-extern "C" {
-
-void init_softmax(bfloat16 max_logit[32], bfloat16 sum_exp[32]) {
-  init_softmax<32>(max_logit, sum_exp);
-}
-
-void online_softmax(bfloat16 attention_score[32][32],
-                    bfloat16 prev_max_logit[32], bfloat16 prev_sum_exp[32],
-                    bfloat16 attention_weight[32][32], bfloat16 scale_exp[32],
-                    bfloat16 new_max_logit[32], bfloat16 new_sum_exp[32]) {
-  const int ROW = 32;
-  const int CHUNK_SIZE = 32;
-  constexpr int vec_factor = 256 / (sizeof(bfloat16) * 8);
-  const int F = CHUNK_SIZE / vec_factor;
-  // ! Note: When using vectorized instructions, the arrays must be properly aligned in memory.
-  alignas(64) bfloat16 tmp_max_logit[32];
-  for (int r = 0; r < ROW; ++r) {
-    bfloat16 *score_row_ptr = &attention_score[r][0];
-    // row max
-    bfloat16 row_max = prev_max_logit[r];
-    for (int i = 0; i < F; i++) {
-      aie::vector<bfloat16, vec_factor> scores =
-          aie::load_v<vec_factor>(score_row_ptr);
-      row_max = std::max(row_max, bfloat16(aie::reduce_max(scores) * 0.125f));
-      score_row_ptr += vec_factor;
-    }
-    // max_logit - new_max
-    tmp_max_logit[r] = bfloat16((prev_max_logit[r] - row_max));
-    new_max_logit[r] = bfloat16(row_max);
-    // logits - new_max
-    score_row_ptr = &attention_score[r][0];
-    for (int i = 0; i < F; i++) {
-      aie::vector<bfloat16, vec_factor> scores =
-          aie::load_v<vec_factor>(score_row_ptr);
-      scores = aie::sub(aie::mul(scores, bfloat16(0.125f)), row_max);
-      aie::store_v(score_row_ptr, scores);
-      score_row_ptr += vec_factor;
-    }
-  }
-  // scale = np.exp(max_logit - new_max)
-  const bfloat16 *__restrict scaleIn = &tmp_max_logit[0];
-  bfloat16 *__restrict scaleExp = &scale_exp[0];
-  exp_bf16<16>(ROW, scaleIn, scaleExp);
-  for (int r = 0; r < ROW; ++r) {
-    // exp_logits = exp(logits - new_max)
-    const bfloat16 *__restrict pIn = &attention_score[r][0];
-    bfloat16 *__restrict pExp = &attention_weight[r][0];
-    float accum_exp_val = exp_bf16<16>(CHUNK_SIZE, pIn, pExp);
-    new_sum_exp[r] = prev_sum_exp[r] * scale_exp[r] + accum_exp_val;
-  }
-}
-}
